fl_implementation/requirements.txt
# Federated Learning
flwr[simulation]>=1.5.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Medical Imaging Dataset
medmnist>=3.0.0

# Simulation
ray>=2.0.0

# Utilities
numpy>=1.24.0,<2
pillow>=9.0.0
tqdm>=4.65.0

# Live dashboard
streamlit>=1.0.0
pandas>=1.5.0
pyzmq>=25.0.0
altair>=5.0.0
